+++ resources/lang/en/validation.php
@@ -3,5 +3,6 @@
 return [
     'in_stock' => 'Insufficient stock levels',
     'max_qty' => 'Unable to add product at this quantity',
-    'min_qty' => 'You must add the minimum required quantity.',
+    'min_qty' => 'You must add at least :min_qty items.',
+    'min_batch' => 'You must add this item in multiples of :min_batch'
 ];