+++ src/Core/Baskets/Validators/BasketValidator.php
@@ -25,10 +25,30 @@ class BasketValidator
 
         $realId = (new ProductVariant)->decodeId($variant['id']);
 
+
         $row = DB::table('product_variants')->select('min_qty')->find($realId);
+
+        $validator->addReplacer('min_quantity', function($message, $attribute, $rule, $parameters) use ($row) {
+            return str_replace([':min_qty'], [$row->min_qty], $message);
+        });
 
         $result = $row->min_qty <= $variant['quantity'];
 
         return $result;
+    }
+
+    public function minBatch($value, $parameters, $basketId, $validator)
+    {
+        $variant = array_get($validator->getData(), str_replace('.quantity', '', $value));
+
+        $realId = (new ProductVariant)->decodeId($variant['id']);
+
+        $row = DB::table('product_variants')->select('min_batch')->find($realId);
+
+        $validator->addReplacer('min_batch', function($message, $attribute, $rule, $parameters) use ($row) {
+            return str_replace([':min_batch'], [$row->min_batch], $message);
+        });
+
+        return ($variant['quantity'] % $row->min_batch) === 0;
     }
 }